feat: add genius act campaign and concentrated all campaign specific info
