+++ .nvmrc
@@ -1,1 +1,1 @@
-v20.10.0
+v20.17.0
