+++ src/components/app/userActionFormEmailCongressperson/constants.ts
@@ -1,6 +1,2 @@
-import { YourPoliticianCategory } from '@/utils/shared/yourPoliticianCategory'
-
 export const ANALYTICS_NAME_USER_ACTION_FORM_EMAIL_CONGRESSPERSON =
   'User Action Form Email Congressperson'
-
-export const EMAIL_FLOW_POLITICIANS_CATEGORY: YourPoliticianCategory = 'senate'
